+++ parser_library/src/diagnostic.h
@@ -696,7 +696,7 @@ struct diagnostic_op
 
     static diagnostic_op warn_CIC001(const range& range);
 
-    static diagnostic_op warn_CIC002(const range& range);
+    static diagnostic_op warn_CIC002(const range& range, std::string_view variable_name);
 
     static diagnostic_op warn_U001_drop_had_no_effect(const range& range, std::string_view arg);
     static diagnostic_op warn_U001_drop_had_no_effect(const range& range, int);